Исправление замечаний в рамках этапа № 2 дипломного проекта
